Fix cyclical imports by not using cyclical imports...
