+++ pyshtools/shclasses/shwindow.py
@@ -16,8 +16,8 @@ import copy as _copy
 from .. import shtools as _shtools
 from ..spectralanalysis import spectrum as _spectrum
 
-from . import shcoeffs
-from . import shgrid
+from .shcoeffsgrid import SHCoeffs
+from .shcoeffsgrid import SHGrid
 
 
 __all__ = ['SHWindow', 'SHWindowCap', 'SHWindowMask']
@@ -342,9 +342,8 @@ class SHWindow(object):
 
         coeffs = self.to_array(itaper, normalization=normalization.lower(),
                                csphase=csphase)
-        return shcoeffs.SHCoeffs.from_array(
-            coeffs, normalization=normalization.lower(),
-            csphase=csphase, copy=False)
+        return SHCoeffs.from_array(coeffs, normalization=normalization.lower(),
+                                   csphase=csphase, copy=False)
 
     def to_shgrid(self, itaper, grid='DH2', zeros=None):
         """
@@ -385,17 +384,17 @@ class SHWindow(object):
         if grid.upper() in ('DH', 'DH1'):
             gridout = _shtools.MakeGridDH(self.to_array(itaper), sampling=1,
                                           norm=1, csphase=1)
-            return shgrid.SHGrid.from_array(gridout, grid='DH', copy=False)
+            return SHGrid.from_array(gridout, grid='DH', copy=False)
         elif grid.upper() == 'DH2':
             gridout = _shtools.MakeGridDH(self.to_array(itaper), sampling=2,
                                           norm=1, csphase=1)
-            return shgrid.SHGrid.from_array(gridout, grid='DH', copy=False)
+            return SHGrid.from_array(gridout, grid='DH', copy=False)
         elif grid.upper() == 'GLQ':
             if zeros is None:
                 zeros, weights = _shtools.SHGLQ(self.lwin)
             gridout = _shtools.MakeGridGLQ(self.to_array(itaper), zeros,
                                            norm=1, csphase=1)
-            return shgrid.SHGrid.from_array(gridout, grid='GLQ', copy=False)
+            return SHGrid.from_array(gridout, grid='GLQ', copy=False)
         else:
             raise ValueError(
                 "grid must be 'DH', 'DH1', 'DH2', or 'GLQ'. " +
